requirements.txt
torch
transformers
datasets
scikit-learn
fastapi
uvicorn
pydantic
pyyaml
pytest
streamlit
pandas
plotly
